pyproject.toml
[project]
name = "travel-agent"
version = "1.0"
description = "Implementation of Travel Agent 3.0 by @samylovma"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Mikhail Samylov", email = "Samylov-Mikhail@yandex.com"},
]
dependencies = [
    "python-telegram-bot>=21.0.1",
    "advanced-alchemy>=0.7.4",
    "asyncpg>=0.29.0",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D", "RUF001",
    # Ruff formatter compatibility
    "W191", "E111", "E114", "E117", "D206", "D300", "Q000", "Q001", "Q002", "Q003", "COM812", "COM819", "ISC001", "ISC002",
    # For now mypy doesn't support PEP 695
    "UP040",
]

[tool.pdm.dev-dependencies]
lint = [
    "ruff>=0.3.3",
    "mypy>=1.9.0",
]
